requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
typing-extensions>=4.9.0 